+++ .github/workflows/test_onnxruntime.yml
@@ -27,6 +27,6 @@ jobs:
       run: |
         pip install .[onnxruntime]
     - name: Test with unittest
-      working-directory: tests/onnxruntime
+      working-directory: tests
       run: |
-        python -m unittest test_onnxruntime
+        python -m unittest discover -s onnxruntime -p 'test_*.py'
